fix: middleware 버그 수정
